+++ lib/build.sh
@@ -142,6 +142,14 @@ install_or_reuse_yarn() {
   fi
 }
 
+copy_profile() {
+  local layer_dir=$1
+  local bp_dir=$2
+
+  mkdir -p "${layer_dir}/profile.d"
+  cp "$bp_dir/profile/WEB_CONCURRENCY.sh" "$layer_dir/profile.d"
+}
+
 write_launch_toml() {
   local build_dir=$1
   local launch_toml=$2