+++ CHANGELOG.md
@@ -3,6 +3,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## main
+- Add profile.d script ([#53](https://github.com/heroku/nodejs-engine-buildpack/pull/53))
 
 ## 0.5.0 (2020-07-16)
 ### Added